move base clases to correct place
